app/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
google-api-python-client
google-auth
google-auth-oauthlib
python-dotenv
